refactor(footer): Add Suspense to Footer Data Featching
Refactor footer adding placeholder styles for loading
