+++ app/data/animeGenres.json
@@ -1,0 +1,12 @@
+[
+  { "name": "Action", "value": "action" },
+  { "name": "Adventure", "value": "adventure" },
+  { "name": "Comedy", "value": "comedy" },
+  { "name": "Drama", "value": "drama" },
+  { "name": "Sci-Fi", "value": "sci-fi" },
+  { "name": "Thriller", "value": "thriller" },
+  { "name": "Romance", "value": "romance" },
+  { "name": "Slice of Life", "value": "slice-of-life" },
+  { "name": "Mystery", "value": "mystery" },
+  { "name": "Sports", "value": "sports" }
+]
